+++ blocksuite/presets/src/effects.ts
@@ -7,8 +7,6 @@ import {
   PageEditor,
 } from './editors/index.js';
 import { CommentInput } from './fragments/comment/comment-input.js';
-import { BacklinkButton } from './fragments/doc-meta-tags/backlink-popover.js';
-import { effects as docTitleEffects } from './fragments/doc-title/index.js';
 import {
   AFFINE_FRAME_PANEL_BODY,
   FramePanelBody,
@@ -70,8 +68,6 @@ import {
 } from './fragments/outline/header/outline-setting-menu.js';
 
 export function effects() {
-  docTitleEffects();
-
   customElements.define('page-editor', PageEditor);
   customElements.define('comment-input', CommentInput);
   customElements.define(
@@ -82,7 +78,6 @@ export function effects() {
   customElements.define(AFFINE_OUTLINE_NOTICE, OutlineNotice);
   customElements.define('comment-panel', CommentPanel);
   customElements.define(AFFINE_OUTLINE_PANEL, OutlinePanel);
-  customElements.define('backlink-button', BacklinkButton);
   customElements.define(AFFINE_OUTLINE_PANEL_HEADER, OutlinePanelHeader);
   customElements.define('affine-editor-container', AffineEditorContainer);
   customElements.define(AFFINE_OUTLINE_NOTE_CARD, OutlineNoteCard);
