--- blocksuite/presets/src/fragments/doc-title/index.ts
@@ -1,1 +1,0 @@
-export * from '@blocksuite/affine-components/doc-title';
