+++ blocksuite/presets/src/fragments/index.ts
@@ -1,4 +1,3 @@
 export * from './comment/index.js';
-export * from './doc-title/index.js';
 export * from './frame-panel/index.js';
 export * from './outline/index.js';
