+++ blocksuite/presets/src/fragments/outline/utils/scroll.ts
@@ -1,9 +1,8 @@
 import type { EditorHost } from '@blocksuite/block-std';
-import { NoteDisplayMode } from '@blocksuite/blocks';
+import { getDocTitleByEditorHost, NoteDisplayMode } from '@blocksuite/blocks';
 import { clamp, DisposableGroup } from '@blocksuite/global/utils';
 
 import type { AffineEditorContainer } from '../../../editors/editor-container.js';
-import { getDocTitleByEditorHost } from '../../doc-title/index.js';
 import { getHeadingBlocksFromDoc } from './query.js';
 
 export function scrollToBlock(editor: AffineEditorContainer, blockId: string) {
