+++ packages/frontend/core/src/blocksuite/blocks/ai-chat-block/ai-chat-spec.ts
