+++ packages/frontend/core/src/blocksuite/blocks/ai-chat-block/components/image-placeholder.ts
