+++ packages/frontend/core/src/blocksuite/blocks/ai-chat-block/model/ai-chat-model.ts
