+++ packages/frontend/core/src/blocksuite/blocks/ai-chat-block/model/consts.ts
