+++ packages/frontend/core/src/blocksuite/blocks/ai-chat-block/model/index.ts
@@ -1,0 +1,3 @@
+export * from './ai-chat-model';
+export * from './consts';
+export * from './types';
