+++ packages/frontend/core/src/blocksuite/blocks/ai-chat-block/model/types.ts
