+++ packages/frontend/core/src/blocksuite/blocks/ai-chat-block/styles.ts
