+++ packages/frontend/core/src/blocksuite/presets/ai/index.ts
@@ -1,3 +1,4 @@
+export * from './_common/config';
 export * from './actions/index';
 export * from './ai-spec';
 export { ChatPanel } from './chat-panel/index';
@@ -6,3 +7,4 @@ export * from './entries/index';
 export * from './messages/index';
 export { AIChatBlockPeekViewTemplate } from './peek-view/chat-block-peek-view';
 export * from './provider';
+export * from './utils/edgeless';
