--- packages/frontend/core/src/blocksuite/presets/blocks/ai-chat-block/index.ts
@@ -1,3 +1,0 @@
-export * from './ai-chat-block.js';
-export * from './ai-chat-edgeless-block.js';
-export * from './ai-chat-spec.js';
