--- packages/frontend/core/src/blocksuite/presets/blocks/index.ts
@@ -1,2 +1,0 @@
-export { AIChatMessages } from './ai-chat-block/components/ai-chat-messages.js';
-export * from './ai-chat-block/index.js';
