+++ packages/frontend/core/src/blocksuite/presets/index.ts
@@ -1,1 +1,2 @@
-export * from './_common/components/text-renderer';
+export * from './_common';
+export * from './ai';
