+++ packages/frontend/core/src/components/blocksuite/block-suite-editor/ai/request.ts
@@ -1,4 +1,4 @@
-import { AIProvider } from '@affine/core/blocksuite/presets/ai';
+import { AIProvider } from '@affine/core/blocksuite/presets';
 import { assertExists } from '@blocksuite/affine/global/utils';
 import { partition } from 'lodash-es';
