+++ packages/frontend/core/src/modules/workspace/global-schema.ts
@@ -1,7 +1,6 @@
+import { AIChatBlockSchema } from '@affine/core/blocksuite/blocks/ai-chat-block/model';
 import { AffineSchemas } from '@blocksuite/affine/blocks/schemas';
 import { Schema } from '@blocksuite/affine/store';
-
-import { AIChatBlockSchema } from '../../blocksuite/blocks';
 
 let _schema: Schema | null = null;
 export function getAFFiNEWorkspaceSchema() {
